minor map_parse restructurization
